Breadcrumbs + adding defects stots + routing fixes

+++ Frescode/DAL/Entities/DefectionSpot.cs
@@ -7,8 +7,8 @@ namespace Frescode.DAL.Entities
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public int OrderNumber { get; set; }
         public ICollection<Picture> AttachedPictures { get; set; } 
         public DateTime DateCreated { get; set; }
